# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  CreateRoomRequest:
    properties:
      description:
        type: string
      floor:
        type: integer
      rent_price:
        type: number
      room_number:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  CreateRoomTypeRequest:
    properties:
      id:
        type: string
      name:
        type: string
      sort_order:
        type: integer
    type: object
  CreateTenantRequest:
    properties:
      email:
        type: string
      emergency_contact:
        type: string
      first_name:
        type: string
      id_card:
        type: string
      last_name:
        type: string
      note:
        type: string
      phone:
        type: string
    type: object
...
  Room:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      dormitory_id:
        type: string
      floor:
        type: integer
      id:
        type: string
      rent_price:
        type: number
      room_number:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      updated_by_name:
        type: string
    type: object
  RoomType:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
...
  Tenant:
    properties:
      active_room_numbers:
        items:
          type: string
        type: array
      created_at:
        type: string
      created_by:
        type: string
      dormitory_id:
        type: string
      email:
        type: string
      emergency_contact:
        type: string
      first_name:
        type: string
      id:
        type: string
      id_card:
        type: string
      last_name:
        type: string
      note:
        type: string
      phone:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      updated_by_name:
        type: string
    type: object
...
  UpdateRoomRequest:
    properties:
      description:
        type: string
      floor:
        type: integer
      rent_price:
        type: number
      room_number:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  UpdateRoomTypeRequest:
    properties:
      name:
        type: string
      sort_order:
        type: integer
    type: object
  UpdateTenantRequest:
    properties:
      email:
        type: string
      emergency_contact:
        type: string
      first_name:
        type: string
      id_card:
        type: string
      last_name:
        type: string
      note:
        type: string
      phone:
        type: string
    type: object
...
info:
  contact: {}
  title: API
paths:
...
  /room-types:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/RoomType'
            type: array
      security:
      - ApiKeyAuth: []
      summary: รายชื่อประเภทห้องพัก
      tags:
      - room-types
    post:
      consumes:
      - application/json
      parameters:
      - description: Room type payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateRoomTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/RoomType'
      security:
      - ApiKeyAuth: []
      summary: สร้างประเภทห้องพัก
      tags:
      - room-types
  /room-types/{id}:
    delete:
      parameters:
      - description: Room Type ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: ลบประเภทห้องพัก
      tags:
      - room-types
    get:
      parameters:
      - description: Room Type ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RoomType'
      security:
      - ApiKeyAuth: []
      summary: ดูข้อมูลประเภทห้องพักตาม ID
      tags:
      - room-types
    put:
      consumes:
      - application/json
      parameters:
      - description: Room Type ID
        in: path
        name: id
        required: true
        type: string
      - description: Room type payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateRoomTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RoomType'
      security:
      - ApiKeyAuth: []
      summary: แก้ไขประเภทห้องพัก
      tags:
      - room-types
  /rooms:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Room'
            type: array
      security:
      - ApiKeyAuth: []
      summary: รายชื่อห้องพัก
      tags:
      - rooms
    post:
      consumes:
      - application/json
      parameters:
      - description: Room payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Room'
      security:
      - ApiKeyAuth: []
      summary: สร้างห้องพัก
      tags:
      - rooms
  /rooms/{id}:
    delete:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: ลบห้องพัก
      tags:
      - rooms
    get:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Room'
      security:
      - ApiKeyAuth: []
      summary: ดูข้อมูลห้องพักตาม ID
      tags:
      - rooms
    put:
      consumes:
      - application/json
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      - description: Room payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Room'
      security:
      - ApiKeyAuth: []
      summary: แก้ไขห้องพัก
      tags:
      - rooms
...
  /tenants:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Tenant'
            type: array
      security:
      - ApiKeyAuth: []
      summary: รายชื่อผู้เช่า
      tags:
      - tenants
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Tenant'
      security:
      - ApiKeyAuth: []
      summary: สร้างผู้เช่า
      tags:
      - tenants
  /tenants/{id}:
    delete:
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: ลบผู้เช่า
      tags:
      - tenants
    get:
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Tenant'
      security:
      - ApiKeyAuth: []
      summary: ดูข้อมูลผู้เช่าตาม ID
      tags:
      - tenants
    put:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Tenant'
      security:
      - ApiKeyAuth: []
      summary: แก้ไขผู้เช่า
      tags:
      - tenants
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"